Vue.JS strongly recommends using kebab-case in templates. Camelcase prop names get converted to kebab-case, preventing dnd-grid from being used with tiles in templates.
This Pull Request changes all props to kebab-case so they can be used as props and dynamic template variables. This results in breaking changes.

An example for the problem, for a template:

```<template>
  <dnd-grid-box :boxId="boxId">
    <h1 v-html="boxId"></h1>
  </dnd-grid-box>
</template>1```

Gets converted to this in the Dom

```<dnd-grid-box box-id="box-1"><h1>box-1</h1></dnd-grid-box>```

Because the prop data attributes gets assigned to box-id, dnd-grid fails to initialize the box.

More info can be found (here)[https://github.com/dattn/dnd-grid]